server/requirements.txt
flask==2.3.2
flask-cors==3.0.10
transformers==4.30.2
torch==2.0.1
pymongo==4.3.3
python-dotenv==1.0.0